Refactor adapters, change di, improve base fragment

+++ domain/src/main/java/com/gmail/maystruks08/domain/interactor/dose/ProductPortionInteractor.kt
@@ -8,5 +8,5 @@ interface ProductPortionInteractor {
 
     fun getProducts(): Single<List<Product>>
 
-    fun onPortionValueChanged(newValue: Int, productId: Int): Completable
+    fun onPortionValueChanged(newValue: Int, productId: Long): Completable
 }
